+++ R/export.R
@@ -18,6 +18,7 @@
 #'     \item \href{https://github.com/csvy}{CSVY} (CSV with a YAML metadata header) using \code{\link[csvy]{write_csvy}}. The YAML header lines are preceded by R comment symbols (\#) by default; this can be turned off by passing a \code{comment_header = FALSE} argument to \code{export}. Setting \code{fwrite = TRUE} (the default) will rely on \code{\link[data.table]{fwrite}} for much faster export.
 #'     \item Pipe-separated data (.psv), using \code{\link[data.table]{fwrite}} or, if \code{fwrite = TRUE}, \code{\link[utils]{write.table}} with \code{sep = '|'} and \code{row.names = FALSE}.
 #'     \item Feather R/Python interchange format (.feather), using \code{feather::write_feather}
+#'     \item Fast storage (.fst), using \code{\link[fst]{write.fst}}
 #'     \item Fixed-width format data (.fwf), using \code{\link[utils]{write.table}} with \code{row.names = FALSE}, \code{quote = FALSE}, and \code{col.names = FALSE}
 #'     \item Serialized R objects (.rds), using \code{\link[base]{saveRDS}}
 #'     \item Saved R objects (.RData,.rda), using \code{\link[base]{save}}
@@ -29,6 +30,7 @@
 #'     \item "XBASE" database files (.dbf), using \code{\link[foreign]{write.dbf}}
 #'     \item Weka Attribute-Relation File Format (.arff), using \code{\link[foreign]{write.arff}}
 #'     \item R syntax object (.R), using \code{\link[base]{dput}} (by default) or \code{\link[base]{dump}} (if \code{format = 'dump'}
+#'     \item Matlab (.mat), using \code{\link[rmatio]{write.mat}}
 #'     \item Excel (.xlsx), using \code{\link[openxlsx]{write.xlsx}}
 #'     \item OpenDocument Spreadsheet (.ods), using \code{\link[readODS]{write_ods}}. (Currently only single-sheet exports are supported.)
 #'     \item XML (.xml), using a custom method based on \code{\link[xml2]{xml_add_child}} to create a simple XML tree and \code{\link[xml2]{write_xml}} to write to disk.